graph classes modules renamed to upper case
